python_weather/requirements.txt
# FarmTech Solutions - Dependências Python
# Sistema de Irrigação Inteligente

# Biblioteca para requisições HTTP (API OpenWeatherMap)
requests>=2.28.0

# Biblioteca para comunicação serial com ESP32
pyserial>=3.5

# Biblioteca para manipulação de dados (opcional - análises)
pandas>=1.5.0

# Biblioteca para visualizações (opcional)
matplotlib>=3.6.0

# Biblioteca para análise numérica (opcional)
numpy>=1.23.0